+++ assert/assertion.go
@@ -6,9 +6,9 @@ import (
 	"github.com/dogmatiq/testkit/render"
 )
 
-// Assertion is a predicate that checks if some specific critiria was met during
-// the execution of a test.
-type Assertion interface {
+// OptionalAssertion is an interface that accept all Assertion types, as well as
+// the Nothing value.
+type OptionalAssertion interface {
 	fact.Observer
 
 	// Begin is called to prepare the assertion for a new test.
@@ -20,7 +20,10 @@ type Assertion interface {
 	End()
 
 	// Ok returns true if the assertion passed.
-	Ok() bool
+	//
+	// If asserted is false, the assertion was a no-op and the value of pass is
+	// meaningless.
+	Ok() (ok bool, asserted bool)
 
 	// BuildReport generates a report about the assertion.
 	//
@@ -28,4 +31,27 @@ type Assertion interface {
 	// the same value as returned from Ok() when this assertion is used as a
 	// sub-assertion inside a composite.
 	BuildReport(ok, verbose bool, r render.Renderer) *Report
+}
+
+// An Assertion is a predicate for determining whether some specific criteria
+// was met during a test.
+type Assertion interface {
+	OptionalAssertion
+
+	// MustOk returns true if the assertion passed.
+	MustOk() bool
+}
+
+// Nothing is an "optional assertion" that always passes and does not build an
+// report.
+var Nothing OptionalAssertion = noopAssertion{}
+
+type noopAssertion struct{}
+
+func (noopAssertion) Notify(fact.Fact)         {}
+func (noopAssertion) Begin(compare.Comparator) {}
+func (noopAssertion) End()                     {}
+func (noopAssertion) Ok() (bool, bool)         { return false, false }
+func (noopAssertion) BuildReport(bool, bool, render.Renderer) *Report {
+	panic("not implemented")
 }