+++ assert/composite_test.go
@@ -6,7 +6,6 @@ import (
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/dogma/fixtures"
 	"github.com/dogmatiq/testkit"
-	"github.com/dogmatiq/testkit/assert"
 	. "github.com/dogmatiq/testkit/assert"
 	"github.com/dogmatiq/testkit/compare"
 	"github.com/dogmatiq/testkit/engine/fact"
@@ -227,15 +226,17 @@ const (
 
 func (a constAssertion) Begin(compare.Comparator) {}
 func (a constAssertion) End()                     {}
-func (a constAssertion) Ok() bool                 { return bool(a) }
+func (a constAssertion) Ok() (bool, bool)         { return bool(a), true }
+func (a constAssertion) MustOk() bool             { return bool(a) }
 func (a constAssertion) Notify(fact.Fact)         {}
-func (a constAssertion) BuildReport(ok, verbose bool, r render.Renderer) *assert.Report {
+
+func (a constAssertion) BuildReport(ok, verbose bool, r render.Renderer) *Report {
 	c := "<always fail>"
 	if a {
 		c = "<always pass>"
 	}
 
-	return &assert.Report{
+	return &Report{
 		TreeOk:   ok,
 		Ok:       bool(a),
 		Criteria: c,