+++ assert/user.go
@@ -71,8 +71,17 @@ func (a *userAssertion) End() {
 }
 
 // Ok returns true if the assertion passed.
-func (a *userAssertion) Ok() bool {
-	return a.s.skipped || !a.s.failed
+//
+// If asserted is false, the assertion was a no-op and the value of pass is
+// meaningless.
+func (a *userAssertion) Ok() (ok bool, asserted bool) {
+	return a.s.skipped || !a.s.failed, true
+}
+
+// MustOk returns true if the assertion passed.
+func (a *userAssertion) MustOk() bool {
+	ok, _ := a.Ok()
+	return ok
 }
 
 // BuildReport generates a report about the assertion.
@@ -83,7 +92,7 @@ func (a *userAssertion) Ok() bool {
 func (a *userAssertion) BuildReport(ok, verbose bool, r render.Renderer) *Report {
 	rep := &Report{
 		TreeOk:   ok,
-		Ok:       a.Ok(),
+		Ok:       a.MustOk(),
 		Criteria: a.s.name,
 	}
 
@@ -93,7 +102,7 @@ func (a *userAssertion) BuildReport(ok, verbose bool, r render.Renderer) *Report
 		rep.Outcome = "the user-defined assertion failed"
 	}
 
-	if !verbose && (ok || a.Ok()) {
+	if !verbose && (ok || a.MustOk()) {
 		return rep
 	}