+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Added
 
+- **[BC]** Add `assert.Assertion.MustOk()`
+- Add `assert.OptionalAssertion` and `assert.Nothing`
 - Add `Test.AdvanceTime()`, `ByDuration()` and `ToTime()`
 
 ### Removed